Move organizationRepos query to resolvers/organizations
